+++ src/compas_fab/backends/pybullet/backend_features/pybullet_forward_kinematics.py
@@ -37,9 +37,9 @@ class PyBulletForwardKinematics(ForwardKinematics):
             The frame in the world's coordinate system (WCF).
         """
         link_name = options.get("link") or robot.get_end_effector_link_name(group)
-        cached_robot = self.client.get_cached_robot(robot)
-        body_id = self.client.get_uid(cached_robot)
-        link_id = self.client._get_link_id_by_name(link_name, cached_robot)
+        cached_robot_model = self.client.get_cached_robot_model(robot)
+        body_id = self.client.get_uid(cached_robot_model)
+        link_id = self.client._get_link_id_by_name(link_name, cached_robot_model)
         self.client.set_robot_configuration(robot, configuration, group)
         frame = self.client._get_link_frame(link_id, body_id)
         if options.get("check_collision"):